+++ runtime/bin/secure_socket_patch.dart
@@ -129,11 +129,6 @@ class SecurityContext {
   static SecurityContext get defaultContext {
     return _SecurityContext.defaultContext;
   }
-
-  @patch
-  static bool get alpnSupported {
-    return _SecurityContext.alpnSupported;
-  }
 }
 
 class _SecurityContext extends NativeFieldWrapperClass1
@@ -179,8 +174,6 @@ class _SecurityContext extends NativeFieldWrapperClass1
   void setClientAuthoritiesBytes(List<int> authCertBytes, {String password})
       native "SecurityContext_SetClientAuthoritiesBytes";
 
-  static bool get alpnSupported => _alpnSupported();
-  static bool _alpnSupported() native "SecurityContext_AlpnSupported";
   void setAlpnProtocols(List<String> protocols, bool isServer) {
     Uint8List encodedProtocols =
         SecurityContext._protocolsToLengthEncoding(protocols);