+++ tests/standalone/io/secure_socket_alpn_test.dart
@@ -125,9 +125,6 @@ void testInvalidArgumentClientConnect(
 }
 
 main() {
-  if (!SecurityContext.alpnSupported) {
-    return 0;
-  }
   final longname256 = 'p' * 256;
   final String longname255 = 'p' * 255;
   final String strangelongname255 = 'ø' + 'p' * 253;